+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
++ PLGOPNS-244: [refactor] Set currency field for each payment method using the same type field of customer groups
+
 ## 3.0.0
 Release date: August, 2020